move toposort to conda.common where it belongs
